5.0

+++ images-to-webp/images-to-webp.php
@@ -3,11 +3,12 @@
 	Plugin Name: Images to WebP
 	Plugin URI: https://www.paypal.me/jakubnovaksl
 	Description: Convert JPG, PNG and GIF images to WEBP, speed up your web
-	Version: 4.9.1
+	Version: 5.0
 	Author: KubiQ
 	Author URI: https://kubiq.sk
 	Text Domain: images-to-webp
 	Domain Path: /languages
+	License: GPLv2 or later 
 */
 
 defined('ABSPATH') || exit;
@@ -16,34 +17,55 @@ class images_to_webp{
 	var $plugin_admin_page;
 	var $settings;
 	var $tab;
-	var $extensions = array( 'jpg', 'jpeg', 'gif', 'png' );
+	var $extensions = [ 'jpg', 'jpeg', 'gif', 'png' ];
 
 	function __construct(){
 		$this->settings = get_site_option('images_to_webp_settings');
 
-		add_action( 'plugins_loaded', array( $this, 'plugins_loaded' ) );
-		add_action( 'admin_menu', array( $this, 'plugin_menu_link' ) );
-		add_action( 'wp_ajax_convert_old_images', array( $this, 'convert_old_images' ) );
-		add_action( 'wp_ajax_itw_subdirectories', array( $this, 'itw_subdirectories' ) );
-		add_action( 'wp_ajax_itw_get_all_subdirectories', array( $this, 'itw_get_all_subdirectories' ) );
-		add_action( 'admin_enqueue_scripts', array( $this, 'admin_enqueue_scripts' ) );
-		add_filter( 'mod_rewrite_rules', array( $this, 'mod_rewrite_rules' ), 77, 1 );
-		add_filter( 'wp_delete_file', array( $this, 'wp_delete_file' ) );
-		add_filter( 'wp_update_attachment_metadata', array( $this, 'wp_update_attachment_metadata' ), 77, 2 );
-		add_action( 'fly_image_created', array( $this, 'fly_images_to_webp' ), 10, 2 );
-		add_action( 'bis_image_created', array( $this, 'bis_images_to_webp' ), 10, 2 );
+		add_action( 'admin_menu', [ $this, 'plugin_menu_link' ] );
+		add_action( 'wp_ajax_itw_convert_old_images', [ $this, 'itw_convert_old_images' ] );
+		add_action( 'wp_ajax_itw_subdirectories', [ $this, 'itw_subdirectories' ] );
+		add_action( 'wp_ajax_itw_get_all_subdirectories', [ $this, 'itw_get_all_subdirectories' ] );
+		add_action( 'admin_enqueue_scripts', [ $this, 'admin_enqueue_scripts' ] );
+		add_filter( 'mod_rewrite_rules', [ $this, 'mod_rewrite_rules' ], 77, 1 );
+		add_filter( 'wp_delete_file', [ $this, 'wp_delete_file' ] );
+		add_filter( 'wp_update_attachment_metadata', [ $this, 'wp_update_attachment_metadata' ], 77, 2 );
+		add_action( 'fly_image_created', [ $this, 'fly_images_to_webp' ], 10, 2 );
+		add_action( 'bis_image_created', [ $this, 'bis_images_to_webp' ], 10, 2 );
+		add_action( 'itw_cron_convert_attachment', [ $this, 'cron_convert_attachment' ] );
 
 		$avif_plugin_file = WP_PLUGIN_DIR . '/images-to-avif/images-to-avif.php';
 		if( ! file_exists( $avif_plugin_file ) ){
-			add_action( 'admin_notices', array( $this, 'admin_notices' ) );
-			add_action( 'wp_ajax_avif_notice_dismissed', array( $this, 'avif_notice_dismissed' ) );
+			add_action( 'admin_notices', [ $this, 'admin_notices' ] );
+			add_action( 'wp_ajax_avif_notice_dismissed', [ $this, 'avif_notice_dismissed' ] );
+		}
+	}
+
+	function cron_convert_attachment( $args ){
+		if( isset( $args['data']['file'], $args['data']['sizes'] ) ){
+			$upload = wp_upload_dir();
+			$path = $upload['basedir'] . '/' . dirname( $args['data']['file'] ) . '/';
+			$sizes = [];
+			$sizes['source'] = $upload['basedir'] . '/' . $args['data']['file'];
+			foreach( $args['data']['sizes'] as $key => $size ){
+				$url = $path . $size['file'];
+				if( in_array( $url, $sizes ) ) continue;
+				$sizes[ $key ] = $url;
+			}
+			$sizes = apply_filters( 'itw_sizes', $sizes, $args['attachment_id'] );
+			foreach( $sizes as $size ){
+				if( ! file_exists( $size . '.webp' ) ){
+					$this->convert_image( $size );
+				}
+			}
 		}
 	}
 
 	function admin_notices(){
 		if( current_user_can('manage_options') && ! get_user_meta( get_current_user_id(), 'avif_notice_dismissed' ) ){
-			if( function_exists('get_current_screen') && isset( get_current_screen()->id ) && in_array( get_current_screen()->id, array( 'dashboard', 'plugins', 'plugin-install', 'upload', 'attachment' ) ) ){ ?>
-				<div class="avif-notice notice notice-info is-dismissible"><p><?php printf( esc_html__( 'There is a new %sImages to AVIF%s plugin that can improve your load time even more and it works perfectly with Images to WebP too.', 'images-to-webp' ), '<a href="' . admin_url( 'plugin-install.php?s=kubiq%20Images%20to%20AVIF&tab=search&type=term' ) . '" target="_blank">', '</a>' ) ?></p></div>
+			if( function_exists('get_current_screen') && isset( get_current_screen()->id ) && in_array( get_current_screen()->id, [ 'dashboard', 'plugins', 'plugin-install', 'upload', 'attachment' ] ) ){
+				/* translators: opening and closing link tag */ ?>
+				<div class="avif-notice notice notice-info is-dismissible"><p><?php printf( esc_html__( 'There is a new %1$sImages to AVIF%2$s plugin that can improve your load time even more and it works perfectly with Images to WebP too.', 'images-to-webp' ), '<a href="' . admin_url('plugin-install.php?s=kubiq%20Images%20to%20AVIF&tab=search&type=term') . '" target="_blank">', '</a>' ) ?></p></div>
 				<script>
 				jQuery(document).ready(function($){
 					$(document).on('click', '.avif-notice .notice-dismiss', function(){
@@ -64,16 +86,12 @@ class images_to_webp{
 	function admin_enqueue_scripts( $hook ){
 		if( $hook == 'media_page_images-to-webp' ){
 			if( isset( $_GET['tab'] ) && $_GET['tab'] == 'convert' ){
-				wp_enqueue_style( 'jstree', plugin_dir_url( __FILE__ ) . 'assets/jstree.min.css', array(), '3.2.1' );
-				wp_enqueue_script( 'jstree', plugin_dir_url( __FILE__ ) . 'assets/jstree.min.js', array('jquery'), '3.2.1' );
-				wp_add_inline_script( 'jstree', 'var transparency_status_message = "' . __( 'Please wait, converting your images is in progress...', 'images-to-webp' ) . '", error_message = "' . __( 'Error {{ERROR}}, trying to continue with missing images... ', 'images-to-webp' ) . '"' );
-				wp_enqueue_script( 'itw_convert', plugin_dir_url( __FILE__ ) . 'assets/convert.js', array('jstree'), 1 );
+				wp_enqueue_style( 'jstree', plugin_dir_url( __FILE__ ) . 'assets/jstree.min.css', [], '3.3.17' );
+				wp_enqueue_script( 'jstree', plugin_dir_url( __FILE__ ) . 'assets/jstree.min.js', ['jquery'], '3.3.17', 1 );
+				wp_add_inline_script( 'jstree', 'var transparency_status_message = "' . esc_html__( 'Please wait, converting your images is in progress...', 'images-to-webp' ) . '", error_message = "' . esc_html__( 'Error {{ERROR}}, trying to continue with missing images... ', 'images-to-webp' ) . '"' );
+				wp_enqueue_script( 'itw_convert', plugin_dir_url( __FILE__ ) . 'assets/convert.js', ['jstree'], 1, 1 );
 			}
 		}
-	}
-
-	function plugins_loaded(){
-		load_plugin_textdomain( 'images-to-webp', FALSE, basename( dirname( __FILE__ ) ) . '/languages/' );
 	}
 
 	function activate(){
@@ -83,15 +101,15 @@ class images_to_webp{
 		update_site_option( 'active_images_to_webp', 1 );
 		// maybe load default settings
 		if( ! $this->settings = get_site_option( 'images_to_webp_settings', 0 ) ){
-			$default_method = array_keys( $methods );
+			$default_method = array_keys( $itw_methods );
 			$default_method = current( $default_method );
-			$default_options = array(
+			$default_options = [
 				'upload_convert' => 1,
 				'extensions' => $this->extensions,
 				'webp_quality' => 85,
 				'method' => $default_method,
 				'delete_originals' => 0,
-			);
+			];
 			update_site_option( 'images_to_webp_settings', $default_options );
 			$this->settings = $default_options;
 		}
@@ -99,6 +117,7 @@ class images_to_webp{
 		flush_rewrite_rules( true );
 		// then test again
 		include_once 'tests/configs.php';
+		new itw_config( $this->generate_mod_rewrite_rules() );
 	}
 
 	function mod_rewrite_rules( $default_rules ){
@@ -135,7 +154,7 @@ class images_to_webp{
 	}
 
 	function filter_plugin_actions( $links, $file ){
-		array_unshift( $links, '<a href="upload.php?page=' . basename( __FILE__ ) . '">' . __('Settings') . '</a>' );
+		array_unshift( $links, '<a href="upload.php?page=' . basename( __FILE__ ) . '">' . esc_html__( 'Settings', 'images-to-webp' ) . '</a>' );
 		return $links;
 	}
 
@@ -146,29 +165,29 @@ class images_to_webp{
 			__( 'Images to WebP', 'images-to-webp' ),
 			'manage_options',
 			basename( __FILE__ ),
-			array( $this, 'admin_options_page' )
+			[ $this, 'admin_options_page' ]
 		);
-		add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), array( $this, 'filter_plugin_actions' ), 10, 2 );
+		add_filter( 'plugin_action_links_' . plugin_basename( __FILE__ ), [ $this, 'filter_plugin_actions' ], 10, 2 );
 	}
 
 	function plugin_admin_tabs( $current = 'general' ){
-		$tabs = array(
+		$tabs = [
 			'general' => __( 'General', 'images-to-webp' ),
 			'convert' => __( 'Convert existing images', 'images-to-webp' ),
-		); ?>
+		]; ?>
 		<h2 class="nav-tab-wrapper">
 		<?php foreach( $tabs as $tab => $name ){ ?>
-			<a class="nav-tab <?php echo $tab == $current ? 'nav-tab-active' : '' ?>" href="?page=<?php echo basename( __FILE__ ) ?>&amp;tab=<?php echo esc_attr( $tab ) ?>"><?php echo esc_html( $name ) ?></a>
+			<a class="nav-tab <?php echo $tab == $current ? 'nav-tab-active' : '' ?>" href="?page=<?php echo esc_attr( basename( __FILE__ ) ) ?>&amp;tab=<?php echo esc_attr( $tab ) ?>"><?php echo esc_html( $name ) ?></a>
 		<?php } ?>
 		</h2><br><?php
 	}
 
 	function admin_options_page(){
 		if( get_current_screen()->id != $this->plugin_admin_page ) return;
-		$this->tab = isset( $_GET['tab'] ) && in_array( $_GET['tab'], array( 'general', 'convert' ) ) ? sanitize_text_field( $_GET['tab'] ) : 'general';
+		$this->tab = isset( $_GET['tab'] ) && in_array( $_GET['tab'], [ 'general', 'convert' ] ) ? sanitize_text_field( $_GET['tab'] ) : 'general';
 		if( isset( $_POST['plugin_sent'] ) ){
 			if( check_admin_referer('itw_general') ){
-				$this->settings = array();
+				$this->settings = [];
 
 				$this->settings['upload_convert'] = intval( $_POST['upload_convert'] );
 				if( $this->settings['upload_convert'] !== 1 ){
@@ -184,7 +203,7 @@ class images_to_webp{
 				}
 				
 				$this->settings['method'] = sanitize_text_field( $_POST['method'] );
-				if( ! in_array( $this->settings['method'], array( 'gd', 'imagick' ) ) ){
+				if( ! in_array( $this->settings['method'], [ 'gd', 'imagick' ] ) ){
 					$this->settings['method'] = '';
 				}
 
@@ -198,11 +217,12 @@ class images_to_webp{
 			}
 		} ?>
 		<div class="wrap">
-			<h2><?php _e( 'Images to WebP', 'images-to-webp' ); ?></h2><?php
+			<h2><?php esc_html_e( 'Images to WebP', 'images-to-webp' ); ?></h2><?php
 			
 			$avif_plugin_file = WP_PLUGIN_DIR . '/images-to-avif/images-to-avif.php';
 			if( ! file_exists( $avif_plugin_file ) ){
-				echo '<div class="notice notice-info"><p>' . sprintf( esc_html__( 'There is a new %sImages to AVIF%s plugin that can improve your load time even more and it works perfectly with this plugin too.', 'images-to-webp' ), '<a href="' . admin_url( 'plugin-install.php?s=kubiq%20Images%20to%20AVIF&tab=search&type=term' ) . '" target="_blank">', '</a>' ) . '</p></div>';
+				/* translators: opening and closing link tag */
+				echo '<div class="notice notice-info"><p>' . sprintf( esc_html__( 'There is a new %1$sImages to AVIF%2$s plugin that can improve your load time even more and it works perfectly with this plugin too.', 'images-to-webp' ), '<a href="' . admin_url('plugin-install.php?s=kubiq%20Images%20to%20AVIF&tab=search&type=term') . '" target="_blank">', '</a>' ) . '</p></div>';
 			}
 
 			$this->plugin_admin_tabs( $this->tab );
@@ -216,15 +236,15 @@ class images_to_webp{
 			$image_extension = pathinfo( $file, PATHINFO_EXTENSION );
 			if( in_array( $image_extension, $this->settings['extensions'] ) ){
 				require_once 'methods/method-' . $this->settings['method'] . '.php';
-				$convert = new webp_converter();
+				$convert = new itw_converter();
 				$response = $convert->convertImage( $file, $this->settings['webp_quality'] );
 				if( $response ){
 					if( $response['size']['after'] >= $response['size']['before'] ){
-						unlink( $response['path'] );
+						wp_delete_file( $response['path'] );
 						return false;
 					}else{
 						if( isset( $this->settings['delete_originals'] ) && $this->settings['delete_originals'] === 1 ){
-							unlink( $file );
+							wp_delete_file( $file );
 						}
 					}
 				}
@@ -235,35 +255,31 @@ class images_to_webp{
 	}
 
 	function itw_subdirectories(){
-		if( defined('DOING_AJAX') && DOING_AJAX ){
-			if( current_user_can('administrator') ){
-				if( check_ajax_referer('itw_convert') ){
-					$ABSPATH = str_replace( '\\', '/', $this->get_abspath() );
+		if( defined('DOING_AJAX') && DOING_AJAX && current_user_can('administrator') && check_ajax_referer('itw_convert') ){
+			$ABSPATH = str_replace( '\\', '/', $this->get_abspath() );
 
-					if( $_REQUEST['folder'] == '#' ){
-						$dir = $ABSPATH;
-					}else{
-						$dir = $ABSPATH . $_REQUEST['folder'] . '/';
-					}
+			if( $_REQUEST['folder'] == '#' ){
+				$dir = $ABSPATH;
+			}else{
+				$dir = $ABSPATH . $_REQUEST['folder'] . '/';
+			}
 
-					$response = array();
-					$directories = scandir( $dir );
-					foreach( $directories as $sub ){
-						if( $sub != '.' && $sub != '..' && is_dir( $dir . $sub ) ){
-							$id = explode( $ABSPATH, $dir . $sub );
-							if( isset( $id[1] ) ){
-								$response[] = array(
-									'id' => esc_attr( $id[1] ),
-									'parent' => esc_attr( $_REQUEST['folder'] ),
-									'text' => esc_html( $sub ),
-									'children' => true
-								);
-							}
-						}
+			$response = [];
+			$directories = scandir( $dir );
+			foreach( $directories as $sub ){
+				if( $sub != '.' && $sub != '..' && is_dir( $dir . $sub ) ){
+					$id = explode( $ABSPATH, $dir . $sub );
+					if( isset( $id[1] ) ){
+						$response[] = [
+							'id' => esc_attr( $id[1] ),
+							'parent' => esc_attr( sanitize_text_field( $_REQUEST['folder'] ) ),
+							'text' => esc_html( $sub ),
+							'children' => true
+						];
 					}
-					wp_send_json( $response );
 				}
 			}
+			wp_send_json( $response );
 		}
 		exit();
 	}
@@ -271,7 +287,7 @@ class images_to_webp{
 	function get_all_subdirectories( $folders, $base = false ){
 		$ABSPATH = $this->get_abspath();
 		if( $base === false ) $base = $ABSPATH;
-		$all_folders = array();
+		$all_folders = [];
 		if( is_array( $folders ) ){
 			foreach( $folders as $folder ){
 				if( $folder != '.' && $folder != '..' ){
@@ -306,52 +322,49 @@ class images_to_webp{
 		exit();
 	}
 
-	function convert_old_images(){
-		if( defined('DOING_AJAX') && DOING_AJAX ){
-			if( current_user_can('administrator') ){
-				if( check_ajax_referer('itw_convert') ){
-					$only_missing = intval( $_POST['only_missing'] );
-					if( $only_missing !== 0 ){
-						$only_missing = 1;
+	function itw_convert_old_images(){
+		if( defined('DOING_AJAX') && DOING_AJAX && current_user_can('administrator') && check_ajax_referer('itw_convert') ){
+			$only_missing = intval( $_POST['only_missing'] );
+			if( $only_missing !== 0 ){
+				$only_missing = 1;
+			}
+			$ABSPATH = str_replace( '\\', '/', $this->get_abspath() );
+			$folder = str_replace( ':\\\\', ':/', $_POST['folder'] );
+			$folder = str_replace( '\\\\', '/', $folder );
+			$folder = preg_replace( '#^' . $ABSPATH . '#', '', $folder );
+			$folder = realpath( $ABSPATH . $folder );
+			if( is_dir( $folder ) ){
+				$secure_path = realpath( $ABSPATH );
+				$secure_path_len = strlen( $secure_path );
+				if( substr( $folder, 0, $secure_path_len ) === $secure_path ){
+					$files = scandir( $folder );
+					$converted = 0;
+
+					$skip_until = false;
+					$webp_processed_file = $folder . '/.webp-processed.json';
+					if( file_exists( $webp_processed_file ) ){
+						$skip_until = trim( file_get_contents( $webp_processed_file ) );
 					}
-					$ABSPATH = str_replace( '\\', '/', $this->get_abspath() );
-					$folder = str_replace( ':\\\\', ':/', $_POST['folder'] );
-					$folder = str_replace( '\\\\', '/', $folder );
-					$folder = preg_replace( '#^' . $ABSPATH . '#', '', $folder );
-					$folder = realpath( $ABSPATH . $folder );
-					if( is_dir( $folder ) ){
-						$secure_path = realpath( $ABSPATH );
-						$secure_path_len = strlen( $secure_path );
-						if( substr( $folder, 0, $secure_path_len ) === $secure_path ){
-							$files = scandir( $folder );
-							$converted = 0;
 
-							$skip_until = false;
-							$webp_processed_file = $folder . '/.webp-processed.json';
-							if( file_exists( $webp_processed_file ) ){
-								$skip_until = trim( file_get_contents( $webp_processed_file ) );
-							}
-
-							foreach( $files as $file ){
-								if( ! $only_missing || ! file_exists( $folder . '/' . $file . '.webp' ) ){
-									if( ! in_array( substr( $file, -5 ), [ '.webp', '.avif' ] ) ){
-										if( $skip_until ){
-											if( $skip_until == $folder . '/' . $file ){
-												$skip_until = false;
-											}
-										}else{
-											$converted += $this->convert_image( $folder . '/' . $file ) ? 1 : 0;
-											file_put_contents( $webp_processed_file, $folder . '/' . $file, LOCK_EX );
-										}
+					foreach( $files as $file ){
+						if( ! $only_missing || ! file_exists( $folder . '/' . $file . '.webp' ) ){
+							if( ! in_array( substr( $file, -5 ), [ '.webp', '.avif' ] ) ){
+								if( $skip_until ){
+									if( $skip_until == $folder . '/' . $file ){
+										$skip_until = false;
 									}
+								}else{
+									$converted += $this->convert_image( $folder . '/' . $file ) ? 1 : 0;
+									file_put_contents( $webp_processed_file, $folder . '/' . $file, LOCK_EX );
 								}
 							}
-
-							wp_delete_file( $webp_processed_file );
-
-							printf( __( '%d converted', 'images-to-webp' ), $converted );
 						}
 					}
+
+					wp_delete_file( $webp_processed_file );
+
+					/* translators: amount of converted files */
+					printf( esc_html__( '%d converted', 'images-to-webp' ), $converted );
 				}
 			}
 		}
@@ -359,31 +372,26 @@ class images_to_webp{
 	}
 
 	function wp_delete_file( $path ){
+		global $wp_filesystem;
+
+		if( ! $wp_filesystem ){
+			require_once ABSPATH . 'wp-admin/includes/file.php';
+			WP_Filesystem();
+		}
+
 		$source = $path . '.webp';
-		if( is_writable( $source ) ) unlink( $source );
+
+		if( $wp_filesystem && $wp_filesystem->is_writable( $source ) ){
+			$wp_filesystem->delete( $source );
+		}
+
 		return $path;
 	}
 
-	function wp_update_attachment_metadata( $data, $attachmentId ){
+	function wp_update_attachment_metadata( $data, $attachment_id ){
 		if( $this->settings['upload_convert'] == 1 ){
-			if( $data && isset( $data['file'] ) && isset( $data['sizes'] ) ){
-				$upload = wp_upload_dir();
-				$path = $upload['basedir'] . '/' . dirname( $data['file'] ) . '/';
-				$sizes = array();
-				$sizes['source'] = $upload['basedir'] . '/' . $data['file'];
-				foreach( $data['sizes'] as $key => $size ){
-					$url = $path . $size['file'];
-					if( in_array( $url, $sizes ) ) continue;
-					$sizes[ $key ] = $url;
-				}
-
-				$sizes = apply_filters( 'itw_sizes', $sizes, $attachmentId );
-
-				foreach( $sizes as $size ){
-					if( ! file_exists( $size . '.webp' ) ){
-						$this->convert_image( $size );
-					}
-				}
+			if( isset( $data['file'], $data['sizes'] ) ){
+				wp_schedule_single_event( time() + 60, 'itw_cron_convert_attachment', [ [ 'data' => $data, 'attachment_id' => $attachment_id ] ] );
 			}
 		}
 		return $data;
@@ -404,7 +412,7 @@ class images_to_webp{
 
 $images_to_webp = new images_to_webp();
 
-register_activation_hook( __FILE__, array( $images_to_webp, 'activate' ) );
+register_activation_hook( __FILE__, [ $images_to_webp, 'activate' ] );
 register_deactivation_hook( __FILE__, 'deactivate_images_to_webp' );
 register_uninstall_hook( __FILE__, 'uninstall_images_to_webp' );
